+++ clouds/databricks/modules/indexing/sql/ST_PARTITIONCENTROID.sql
@@ -1,0 +1,1 @@
+CREATE OR REPLACE FUNCTION ST_PARTITIONCENTROID as 'com.carto.analyticstoolbox.modules.index.ST_PartitionCentroid';