+++ clouds/databricks/modules/indexing/sql/ST_Z2LATLON.sql
@@ -1,0 +1,1 @@
+CREATE OR REPLACE FUNCTION ST_Z2LATLON as 'com.carto.analyticstoolbox.modules.index.ST_Z2LatLon';