+++ clouds/databricks/modules/transformations/CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.0.1] - 2022-08-29
+
+- Add ST_BUFFERPOINT function.
+- Add ST_DIFFERENCE function.
+- Add ST_SIMPLIFY function.
+- Add ST_SIMPLYFYPRESERVETOPOLOGY function.
+
 ## [1.0.0] - 2022-08-19
 
 ### Added